Add view for handling next_location_popup
The view returns the HTML form that will be
embedded in the popup if the request is a GET
request, otherwise it serves as an endpoint for
posting new data to.

Todo: Figure out a better way to select datetime.
As it is now, any time not in the current day will
not be properly displayed.